refactor(desktop): centralize chat permissions
